requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
validators>=0.20.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pydantic==2.5.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0

# ML and embeddings
sentence-transformers==2.2.2
transformers==4.36.0
torch>=2.1.0
scikit-learn>=1.3.2

# Vector database
chromadb>=1.0.0

# LLM integration
huggingface-hub==0.19.4
accelerate==0.25.0

# Web framework
streamlit==1.28.1
streamlit-chat==0.1.1

# Database
pymongo==4.6.0

# Utilities
tqdm==4.66.1
loguru==0.7.2
pytest==7.4.3
python-dateutil==2.8.2

# Text processing
nltk==3.8.1
spacy
langdetect==1.0.9

# HTTP and async
httpx==0.25.2
aiohttp==3.9.1 

# Linting and formatting
ruff==0.4.5

# Static type checking
mypy==1.10.0

# Type stubs for popular packages (optional but useful)
types-requests
types-PyYAML
types-python-dateutil
